pyproject.toml
[tool.poetry]
name = "scraping"
version = "0.1.0"
description = "repository for learning async, requests"
authors = ["luculliano <vederecento@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = {git = "https://github.com/psf/requests.git", rev = "v2.28.2"}
faker = "^18.4.0"
aiohttp = "^3.8.4"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.2"
aiofiles = "^23.1.0"
aiogram = {git = "https://github.com/aiogram/aiogram.git", rev = "v3.0.0b7"}
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
